Updating Version() test to assert it is updated
For the v7.2.3 release, the autorest.Version() function didn't get
updated to reflect the new release. To ensure that doesn't happen again,
the TestVersion() function was updated to make sure that when commited,
the version is greater than the most recently relesed version.

This added a tight coupling with Git and using tags as release
mechanisms.