Updating example-msi-app vendor its dependencies.
example-msi-app is an independent package the currently depends on the
github.com/Azure/go-autorest being in the dev branch. It also declares a
dependency on github.com/Azure/azure-sdk-for-go/arm/resources/resouces,
which in turn also declares a dependency on
github.com/Azure/go-autorest. However, resources relies on the master
branch of go-autorest. In the past that didn't cause conflict, because
one could fast-forward from master to dev and Git just hid the
difference allowing builds to succeed in Travis. However, once v7.3.1
was released, master and dev diverged, causing Git to no longer be able
to fast-forward. This caused a dependency conflict, and Travis builds
began failing. To fix this, I've updated the example-msi-app to use
glide to vendor its dependencies. This allows separation between the
needs of the example app and the rest of the package. However, it does
mean that temporarily autorest/adal/example-msi-app/glide.yaml must
point to non-release commit of go-autorest. Once v8.0.0 is released,
that will be fixed.