Updating JSON Roundtrip for tolerance
Previously, JSONRoundTrip asserted that we are able to have granularity
to nanoseconds even when storing time as a float64 of nanoseconds. We
have run into some trouble with that, and must accept that if one needs
more precision, marshalling as Text or Binary will have to suffice over
encoding it as a JSON number.