Add support for bearer authentication callbacks.
Creation of bearer authorization requires the tenant ID and resource URL
and in some cases it can be tricky to obtain the correct values leading
to authentication failures.  To simplify this we can rely on the bearer
authorization header to obtain the correct tenant ID and resource URL,
providing these bits of data via an authorization callback that's
invoked when a request is submitted.