Add WaitForCompletion to Future type.
The WaitForCompletion method provides a default implementation that will
wait until a future has completed, is cancelled or times out.
Fixed a bug in Future.Done() to not update polling status if the HTTP
response isn't one of the expected values.
Added support to mock framework for returning errors in responses so we
can simulate things like transient network failures.