Future state machine rewrite.
The polling mechanism for async operations is determined by the verb of
the originating request along with the HTTP status code of the initial
response; the original implementation didn't take this into account
which caused some Future types to incorrectly poll or obtain the wrong
URL for retrieving the result.
This change introduces the pollingTracker interface as an abstraction
over verb-specific polling trackers; this allows the logic to be
separated per tracker.  All shared state and methods are in the
pollingTrackerBase type.