Refactor auth package to expose settings
The loading of environment and file settings has been refactored and
exported so that consumers have access to the raw data.
The steps performed in NewAuthorizerFromEnvironment() have been broken
down into methods on EnvironmentSettings so that consumers can build
custom authorization chains.
Added support for certificate authorization for file-based config.