Fix GetTokenFromCLI cannot work in zsh
GetTokenFromCLI uses shell to launch az when obtaining tokens.
This approach may produce an error under zsh as zsh doesn't always
assume the shell script to be located in /usr/bin.

As the environment variable PATH is set when executing, it can
rely on the shell to locate az CLI script rather than explicitly
specifying the shell.