Walk error chain when checking for TokenRefreshError
The refresh methods on MultiTenantServicePrincipalToken didn't properly
wrap the inner error, so as a result the check for TokenRefreshError
would fail leading to retries on a credential that would always fail.
Updated the methods to use %w to wrap the error.
Added Unwrap() method to DetailedError
Updated IsTokenRefreshError() to use errors.As() to walk the error
chain.
Legacy versions for consumers on earlier versions of Go have been
provided.  In these implementations the TokenRefreshError isn't wrapped
to avoid the hang at a loss of error information.