Cache last Retry-After value when polling on LRO
Some operations return a retry-after value on the first response but not
subsequent responses.  This causes the polling logic to fall back to the
default polling value, causing some LROs to take longer than required.
Cache the last retry-after value and use it when no retry-after value is
provided by the service.