Remove GoDep files
Go Dep does not support versioned module paths, so we may as well
remove it.

Note that I did not add a `go.mod` at the root of the repository.
The roof ot this repository is an empty module, and adding a go.mod
will require a major version update, to adjust the name to match the
current release (v14).

Given that (from the README), the "main" tags were intended for projects
that have not yet migrated to modules, adding a go.mod will not benefit
those consumers.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>