Don't raise an error from Client.AddToUserAgent
This is essentially just string concatenation, there is no real reason
to return an error here. This causes clients to write dozens of lines
of pointless error handling or just not handle the err.

Thinking about the contract of this function, if I've asked for it to
append "" (empty string) to the user agent, and it does nothing, then
it has been a success. Raising an error in that case is a bit weird.

`"foo" + ""` does not result in an error, which is essentially what
this is doing.